Change text_layout and font to use floating-point.
This should increase accuracy, as metric::left and metric::top are now
floating-point. This is then passed on to text_layout, which has been
modified to use floating-point calculations.

+++ foundation/visual/font.hpp
@@ -28,8 +28,8 @@ private:
 //--- metric ---------------------------------------------------------------------------------------
 struct metric
 {
-    s32 left;
-    s32 top;
+    f32 left;
+    f32 top;
     f32 advance;
     u16 width;
     u16 height;
